refactor: extract MattermostBridge to its own file and clarify try-block order
